Q1/requirements.txt
uvicorn == 0.20.0
fastapi == 0.90.0
opencv-contrib-python == 4.7.0.68
opencv-python == 4.7.0.68
numpy == 1.21.2
onnxruntime-gpu == 1.12.1
websockets == 10.4
asyncio == 3.4.3
Pillow == 9.0.1